T_defs_pcu: Set default val for X2000 to 0 ms
That timer is really only useful to free the tbf asynchronously after
generating the Pkt Access Reject message, since we have nothing to do
with it after the message is sent, and the dummt TBF doesn't really hold
any reserved resource such as USF or TFI.
The timer is useful to still do the freeing asyncrhonously, since the
scheduler is interacting with the TBF during the code path, but there's
no real need to keep the object alive for 2 ms afterwards. Having a
default value of 0 ms is enough, since it fullfills the requirement
of freeing asnchronously.

The value of 2 ms was set initially when the reject support was added
here (e9a138e11111f509f988807bbdc5ca8cce2b3d3d), with no specific
explanation on the 2 ms value. It was just probably picked as a
convinience one, but 0 is actually more convinient.

Change-Id: I60e34e643f5c9d9afaf85530c54ab3232dc8f0be