Error trying to obtain FN from RFN if curr_fn not known
This may happen if a RACH.ind is received before any DATA.ind has been
received.
With usual osmo-bts-trx or osmo-bts-sysmo, this shouldn't happen
nowadays, but it is still a problem with osmo-bts-virtual, where lower
layers don't submit NOPE.ind in the absence of data, and hence it won't
sent DATA.ind to osmo-pcu all the time.

This change helps in showcasing confusing scenarios where the RFN
generated in the Imm Ass was wrong.

Change-Id: I29b7ba828fe890f90e35686bbb04d4abfe56b955