tbf_dl_fsm: Remove T3169 references, simplify T3195 set up as consequence
T3169 is triggered as consequence of N3101 and N3103 reaching MAX. Those
counters are related to UL TBF only, and hence T3169 also only applies
to UL TBFs. The references to T3169 were there as a remains from time
where we had a single tbf_fsm for both UL and DL TBFs. It can now be
further simplified.

Change-Id: I02f7654518da617ce6704d807d34761dbadecf07