tbf_ul_fsm: Remove 3195 references, simplify T3169 set up as consequence
T3195 is triggered as consequence of N3105 reaching MAX, which only
occurs in DL TBFs. Hence, T3195 also only applies to DL TBFs.
The references to T3195 were there as a remains from time
where we had a single tbf_fsm for both UL and DL TBFs. It can now be
further simplified.

Change-Id: I07a43c13289d50707115187a3f22df21443a7b4a