Abort UL TBF PACCH Ass based on T3168
It makes no sense to continue trying to assign the UL TBF over PACCH
after T3168 * 4 retrans time out.

This helps in releasing the TBF after we got rid of incorrect use of
N3015 in UL TBFs.

While at it, update tbf_ul_fsm to use T3168 instead of X2001, since it
really needs to match T3168. Ideally it would not even have a timer
itself and receive an event from tbf_ul_ass_fsm, but that's left as a
TODO (it was already before) and simply the timer is updated.

Change-Id: I87dff68dedd06b60501e7586d20faf02bb1f0c93