tbf_dl_fsm: Drop impossible event
When sending the last DL block, the FSM moves FLOW->FINISHED. Hence, it
is impossible to receive the FINAL_ACK in state flow, when we didn't yet
sent the last DL block (it's only possible if there's a bug in the MS).

TbfTest need to be adapted since they where acting wrong.

Change-Id: I3288743ff01ff84c3d345b6efb7c3fb6ca934791