bts: get rid of fn parameter in bts_rcv_imm_ass_cnf
The function bts_rcv_imm_ass_cnf, which processes the confirmation
message from osmo-bts or osmo-bsc has an fn (frame number) parameter.
This parameter not used for anything other than logging. The source of
this frame number parameter is the conformation message that comes from
osmo-pcu or osmo-bsc. In the case of osmo-bsc the frame number is always
0, since osmo-bsc uses RSL and can not know the exact frame numbers.

since we do not really need the frame number for the confirmation, lets
remove the fn parameter entirely.

Related: OS#5927
Change-Id: I35bc99eaec5d0287ae3916bc668f0babaddfd6ce