bts: rename struct member gen_idle_blocks to gen_idle_blocks_C0
The struct member gen_idle_blocks controls whether the the PCU has to
generate dummy blocks to fill idle periods or if this task is left to
the BTS hardware.

Since idle dummy blocks only have to be transmitted on the C0 carrier,
we should rename the gen_idle_blocks struct member to
gen_idle_blocks_C0.

Related: OS#6198
Change-Id: If9511ed911f9c5d6d463881a387a9e7d11a15465