fix license headers
I'm not sure why some files did contain an AGPLv3+ header, instead of
the GPLv2+ which is the actual overall project license.  I consider it a
mistake.

In any case, any copyrightable contribution to those files was done by
sysmocom employees, so I as managing directory can legally make a
license change, whther or not it was a mistake early on or not.

The only  remaining AGPLv3+ files are those implementing the "direct
PHY" interface for osmo-bts-{symo,lc15,oc2g}.  That is als
unintentional, but more complex to resolve as they were copied from
osmo-bts, which is indeed an AGPLv3+ project.

Change-Id: I52533193de0ec7e24bf7c5bb035589ba5b9419e4