pcuif_sock: do not retry when sending PCU version
When some client connects to the PCU socket, the PCU sends a TXT
indication with its version number in it. There is a retry mechanism
in place that checks if the BTS got active. In case the BTS is not
active yet, the TXT indication with the version number is repeated.

This mechanism is unnecessary. It even can cause annoyance in cas the
BTS on the other end does not send an info indication because it does
not use GPRS/EGPRS at all. The TXT indication is sent after the PCU
socket connection is made. We use a normal unix domain socket, which
can not lose messages. It is guaranteed that the TXT indication
reaches the client (BTS or BSC).

Related: OS#6270
Change-Id: Iddd3eed113ed288e629b6029f9a56de427de8126